+++ src/NBatch.Core/ItemReader/IReader.cs
@@ -5,6 +5,6 @@ namespace NBatch.Core.ItemReader
     public interface IReader<out TInput>
     {
         int LinesToSkip { get; set; }
-        IEnumerable<TInput> Read(int startIndex, int chunkSize);
+        IEnumerable<TInput> Read(long startIndex, int chunkSize);
     }
 }
