+++ src/NBatch.Core/SkipContext.cs
@@ -4,11 +4,11 @@ namespace NBatch.Core
 {
     public sealed class SkipContext
     {
-        public int LineNumber { get; private set; }
+        public long LineNumber { get; private set; }
         public string StepName { get; set; }
         public Exception Exception { get; private set; }
 
-        public SkipContext(string stepName, int lineNumber, Exception ex)
+        public SkipContext(string stepName, long lineNumber, Exception ex)
         {
             StepName = stepName;
             LineNumber = lineNumber;
